+++ client/keys/add.go
@@ -13,6 +13,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/cmd/gaia/app"
 	"github.com/cosmos/cosmos-sdk/crypto/keys"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/rest"
 
 	"errors"
 
@@ -77,7 +78,7 @@ the flag --nosort is set.
 	cmd.Flags().Bool(flagNoBackup, false, "Don't print out seed phrase (if others are watching the terminal)")
 	cmd.Flags().Bool(flagDryRun, false, "Perform action, but don't add key to local keystore")
 	cmd.Flags().Uint32(flagAccount, 0, "Account number for HD derivation")
-	cmd.Flags().Uint32(flagIndex, 0, "Index number for HD derivation")
+	cmd.Flags().Uint32(flagIndex, 0, "Address index number for HD derivation")
 	return cmd
 }
 
@@ -104,7 +105,7 @@ func runAddCmd(_ *cobra.Command, args []string) error {
 	if viper.GetBool(flagDryRun) {
 		// we throw this away, so don't enforce args,
 		// we want to get a new random seed phrase quickly
-		kb = client.MockKeyBase()
+		kb = keys.NewInMemory()
 		encryptPassword = app.DefaultKeyPass
 	} else {
 		kb, err = NewKeyBaseFromHomeFlag()
@@ -309,7 +310,7 @@ func printCreate(info keys.Info, showMnemonic bool, mnemonic string) error {
 
 // function to just create a new seed to display in the UI before actually persisting it in the keybase
 func generateMnemonic(algo keys.SigningAlgo) string {
-	kb := client.MockKeyBase()
+	kb := keys.NewInMemory()
 	pass := app.DefaultKeyPass
 	name := "inmemorykey"
 	_, seed, _ := kb.CreateMnemonic(name, keys.English, pass, algo)
@@ -399,7 +400,7 @@ func AddNewKeyRequestHandler(indent bool) http.HandlerFunc {
 
 		keyOutput.Mnemonic = mnemonic
 
-		PostProcessResponse(w, cdc, keyOutput, indent)
+		rest.PostProcessResponse(w, cdc, keyOutput, indent)
 	}
 }
 
@@ -488,6 +489,6 @@ func RecoverRequestHandler(indent bool) http.HandlerFunc {
 			return
 		}
 
-		PostProcessResponse(w, cdc, keyOutput, indent)
+		rest.PostProcessResponse(w, cdc, keyOutput, indent)
 	}
 }