+++ x/gov/simulation/invariants.go
@@ -2,11 +2,10 @@ package simulation
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/mock/simulation"
 )
 
 // AllInvariants tests all governance invariants
-func AllInvariants() simulation.Invariant {
+func AllInvariants() sdk.Invariant {
 	return func(ctx sdk.Context) error {
 		// TODO Add some invariants!
 		// Checking proposal queues, no passed-but-unexecuted proposals, etc.