+++ x/staking/genesis_test.go
@@ -20,8 +20,8 @@ func TestInitGenesis(t *testing.T) {
 	ctx, _, keeper := keep.CreateTestInput(t, false, 1000)
 
 	pool := keeper.GetPool(ctx)
-	pool.BondedTokens = TokensFromTendermintPower(2)
-	valTokens := TokensFromTendermintPower(1)
+	pool.BondedTokens = sdk.TokensFromTendermintPower(2)
+	valTokens := sdk.TokensFromTendermintPower(1)
 
 	params := keeper.GetParams(ctx)
 	validators := make([]Validator, 2)
@@ -48,7 +48,7 @@ func TestInitGenesis(t *testing.T) {
 	actualGenesis := ExportGenesis(ctx, keeper)
 	require.Equal(t, genesisState.Pool, actualGenesis.Pool)
 	require.Equal(t, genesisState.Params, actualGenesis.Params)
-	require.Equal(t, genesisState.Bonds, actualGenesis.Bonds)
+	require.Equal(t, genesisState.Delegations, actualGenesis.Delegations)
 	require.EqualValues(t, keeper.GetAllValidators(ctx), actualGenesis.Validators)
 
 	// now make sure the validators are bonded and intra-tx counters are correct
@@ -76,7 +76,7 @@ func TestInitGenesisLargeValidatorSet(t *testing.T) {
 
 	// Assigning 2 to the first 100 vals, 1 to the rest
 	pool := keeper.GetPool(ctx)
-	bondedTokens := TokensFromTendermintPower(int64(200 + (size - 100)))
+	bondedTokens := sdk.TokensFromTendermintPower(int64(200 + (size - 100)))
 	pool.BondedTokens = bondedTokens
 
 	params := keeper.GetParams(ctx)
@@ -89,9 +89,9 @@ func TestInitGenesisLargeValidatorSet(t *testing.T) {
 
 		validators[i].Status = sdk.Bonded
 
-		tokens := TokensFromTendermintPower(1)
+		tokens := sdk.TokensFromTendermintPower(1)
 		if i < 100 {
-			tokens = TokensFromTendermintPower(2)
+			tokens = sdk.TokensFromTendermintPower(2)
 		}
 		validators[i].Tokens = tokens
 		validators[i].DelegatorShares = sdk.NewDecFromInt(tokens)